fix instructions in macros lesson
The previous instructions don't work because the cw command leaves the cursor at the end of DONE, so the j command leaves it on the final character of TODO. The next cw command only changes this final character rather than the whole word.

This adds a ^ command after the j to make it repeatable. It also clarifies the second step which says to "move back" even though the cursor should be on the second line already.